requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
future==0.18.2
idna==2.10
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
serializers==0.2.4
sqlparse==0.4.1
urllib3==1.26.3
